+++ mjolnir/forcefield/global/GlobalPairExcludedVolumeInteraction.hpp
@@ -1,7 +1,7 @@
 #ifndef MJOLNIR_INTERACTION_GLOBAL_PAIR_EXCLUDED_VOLUME_INTEARACTION_HPP
 #define MJOLNIR_INTERACTION_GLOBAL_PAIR_EXCLUDED_VOLUME_INTEARACTION_HPP
-#include <mjolnir/interaction/global/GlobalPairInteraction.hpp>
-#include <mjolnir/potential/global/ExcludedVolumePotential.hpp>
+#include <mjolnir/forcefield/global/GlobalPairInteraction.hpp>
+#include <mjolnir/forcefield/global/ExcludedVolumePotential.hpp>
 #include <mjolnir/core/SimulatorTraits.hpp>
 #include <memory>
