+++ mjolnir/forcefield/global/GlobalPairInteraction.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/interaction/global/GlobalPairInteraction.hpp>
+#include <mjolnir/forcefield/global/GlobalPairInteraction.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
