+++ mjolnir/forcefield/global/GlobalPairUniformLennardJonesInteraction.cpp
@@ -1,4 +1,4 @@
-#include <mjolnir/interaction/global/GlobalPairUniformLennardJonesInteraction.hpp>
+#include <mjolnir/forcefield/global/GlobalPairUniformLennardJonesInteraction.hpp>
 
 #ifndef MJOLNIR_SEPARATE_BUILD
 #error "MJOLNIR_SEPARATE_BUILD flag is required"
